fix: split list routes into layout+index to fix detail view navigation
Library, Authors, Series, and Jobs list routes were implicitly acting as
layout routes (because they had child detail routes), but did not render
<Outlet />, so navigating to a detail URL updated the address bar and ran
the loader but never mounted the detail component.

Split each affected route into a thin layout (with <Outlet />) and an index
route (with list page content), matching the existing settings.tsx pattern.

Closes #67
