+++ apps/web/src/routes/_authenticated/-library.test.tsx
@@ -1,786 +1,28 @@
 // @vitest-environment happy-dom
-import type * as DataTableModule from "~/components/data-table";
 import type * as TanstackRouter from "@tanstack/react-router";
 import { render, screen } from "@testing-library/react";
-import { describe, it, expect, vi, beforeEach } from "vitest";
-
-const defaultFacetCounts = {
-  format: [
-    { formatFamily: "EBOOK", _count: { _all: 1 } },
-  ],
-  hasCover: { withCover: 0, withoutCover: 1 },
-  series: 0,
-};
-
-let mockLoaderData: {
-  libraryResult: {
-    works: {
-      id: string;
-      titleDisplay: string;
-      titleCanonical: string;
-      sortTitle: string;
-      coverPath: string | null;
-      createdAt: Date;
-      enrichmentStatus: string;
-      series: { id: string; name: string } | null;
-      seriesPosition: number | null;
-      editions: {
-        formatFamily: string;
-        publisher: string;
-        isbn13: string | null;
-        isbn10: string | null;
-        contributors: { role: string; contributor: { nameDisplay: string } }[];
-      }[];
-    }[];
-    totalCount: number;
-    facetCounts: typeof defaultFacetCounts;
-  };
-  activeJobCount: number;
-  progressMap: Record<string, number>;
-} = {
-  libraryResult: {
-    works: [],
-    totalCount: 0,
-    facetCounts: defaultFacetCounts,
-  },
-  activeJobCount: 0,
-  progressMap: {},
-};
-
-let mockSearch = { page: 1, pageSize: 50, sort: "title-asc" as const };
-const mockNavigate = vi.fn().mockImplementation((opts: { search?: (prev: Record<string, unknown>) => unknown }) => {
-  if (typeof opts.search === "function") {
-    opts.search({});
-  }
-});
+import { describe, it, expect, vi } from "vitest";
 
 vi.mock("@tanstack/react-router", async () => {
   const actual = await vi.importActual<typeof TanstackRouter>("@tanstack/react-router");
   return {
     ...actual,
-    Link: ({ children, to, params, ...props }: { children?: React.ReactNode; to: string; params?: Record<string, string>; [key: string]: unknown }) => {
-      let href = to;
-      if (params) {
-        for (const [key, value] of Object.entries(params)) {
-          href = href.replace(`$${key}`, value);
-        }
-      }
-      return <a href={href} {...props}>{children}</a>;
-    },
+    Outlet: () => <div data-testid="outlet">outlet</div>,
     useRouter: () => ({ invalidate: vi.fn(), navigate: vi.fn() }),
-    useNavigate: () => mockNavigate,
-    createFileRoute: (_path: string) => (opts: Record<string, unknown>) => {
-      // Call validateSearch and loaderDeps to exercise those branches
-      if (typeof opts.validateSearch === "function") {
-        (opts.validateSearch as (s: unknown) => unknown)({});
-      }
-      if (typeof opts.loaderDeps === "function") {
-        (opts.loaderDeps as (s: { search: unknown }) => unknown)({ search: {} });
-      }
-      return {
-        ...opts,
-        options: opts,
-        useLoaderData: () => mockLoaderData,
-        useSearch: () => mockSearch,
-        useRouteContext: () => ({}),
-      };
-    },
+    createFileRoute: (_path: string) => (opts: Record<string, unknown>) => ({
+      ...opts,
+      options: opts,
+      useLoaderData: () => ({}),
+      useRouteContext: () => ({}),
+    }),
   };
 });
 
-vi.mock("~/hooks/use-sse", () => ({
-  useSSE: vi.fn(),
-}));
-
-const getFilteredLibraryWorksServerFnMock = vi.fn();
-vi.mock("~/lib/server-fns/library", () => ({
-  getFilteredLibraryWorksServerFn: getFilteredLibraryWorksServerFnMock,
-}));
-
-const getActiveJobCountServerFnMock = vi.fn();
-vi.mock("~/lib/server-fns/import-jobs", () => ({
-  getActiveJobCountServerFn: getActiveJobCountServerFnMock,
-}));
-
-const getBulkReadingProgressServerFnMock = vi.fn();
-vi.mock("~/lib/server-fns/reading-progress", () => ({
-  getBulkReadingProgressServerFn: getBulkReadingProgressServerFnMock,
-}));
-
-let mockView = "grid" as "grid" | "table";
-const mockSetView = vi.fn();
-vi.mock("~/hooks/use-library-view-preference", () => ({
-  useLibraryViewPreference: () => [mockView, mockSetView],
-}));
-
-vi.mock("~/components/library-grid", () => ({
-  LibraryGrid: ({ works, progressMap }: { works: unknown[]; progressMap?: Record<string, number> }) => (
-    <div data-testid="library-grid" data-progress-map={progressMap ? JSON.stringify(progressMap) : undefined}>Grid: {String(works.length)} works</div>
-  ),
-}));
-
-let capturedToolbarProps: Record<string, unknown> = {};
-vi.mock("~/components/library-toolbar", () => ({
-  LibraryToolbar: (props: Record<string, unknown>) => {
-    capturedToolbarProps = props;
-    return (
-      <div data-testid="library-toolbar" data-view={props.view as string} data-filter={props.filterValue as string} />
-    );
-  },
-}));
-
-let capturedFiltersProps: Record<string, unknown> = {};
-vi.mock("~/components/library-filters", () => ({
-  LibraryFilters: (props: Record<string, unknown>) => {
-    capturedFiltersProps = props;
-    return (
-      <div data-testid="library-filters" data-filters={JSON.stringify(props.filters)} />
-    );
-  },
-}));
-
-let capturedPaginationProps: Record<string, unknown> = {};
-vi.mock("~/components/library-pagination", () => ({
-  LibraryPagination: (props: Record<string, unknown>) => {
-    capturedPaginationProps = props;
-    return (
-      <div data-testid="library-pagination" data-page={String(props.page)} data-total={String(props.totalCount)} />
-    );
-  },
-}));
-
-// Use real data-table so column cell renderers execute
-vi.mock("~/components/data-table", async () => {
-  const actual = await vi.importActual<typeof DataTableModule>("~/components/data-table");
-  return actual;
-});
-
-vi.mock("~/components/skeletons/grid-page-skeleton", () => ({
-  GridPageSkeleton: () => <div>Loading grid...</div>,
-}));
-
-vi.mock("@tanstack/react-virtual", () => ({
-  useVirtualizer: ({ count }: { count: number }) => ({
-    getVirtualItems: () =>
-      count > 0
-        ? Array.from({ length: Math.min(count, 5) }, (_, i) => ({
-            index: i,
-            start: i * 48,
-            end: (i + 1) * 48,
-          }))
-        : [],
-    getTotalSize: () => count * 48,
-  }),
-}));
-
-vi.mock("~/lib/library-search-schema", () => ({
-  librarySearchSchema: { parse: (v: unknown) => v },
-}));
-
-const makeWork = (title: string, authors: string[] = [], formats: string[] = [], enrichmentStatus = "ENRICHED") => ({
-  id: `work-${title.toLowerCase().replace(/\s/g, "-")}`,
-  titleDisplay: title,
-  titleCanonical: title.toLowerCase(),
-  sortTitle: title.toLowerCase(),
-  coverPath: null,
-  createdAt: new Date("2025-01-01"),
-  enrichmentStatus,
-  series: null,
-  seriesPosition: null,
-  editions: [
-    {
-      formatFamily: formats[0] ?? "EBOOK",
-      publisher: "Test Publisher",
-      isbn13: "1234567890123",
-      isbn10: null,
-      contributors: authors.map((name) => ({
-        role: "AUTHOR",
-        contributor: { nameDisplay: name },
-      })),
-    },
-  ],
-});
-
-describe("LibraryPage", () => {
-  beforeEach(() => {
-    mockLoaderData = {
-      libraryResult: {
-        works: [],
-        totalCount: 0,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    mockSearch = { page: 1, pageSize: 50, sort: "title-asc" };
-    mockView = "grid";
-    capturedToolbarProps = {};
-    capturedFiltersProps = {};
-    capturedPaginationProps = {};
-    vi.clearAllMocks();
-  });
-
-  it("loader calls getFilteredLibraryWorksServerFn, getActiveJobCountServerFn, and getBulkReadingProgressServerFn", async () => {
-    getFilteredLibraryWorksServerFnMock.mockResolvedValueOnce({ works: [], totalCount: 0, facetCounts: defaultFacetCounts });
-    getActiveJobCountServerFnMock.mockResolvedValueOnce(0);
-    getBulkReadingProgressServerFnMock.mockResolvedValueOnce({});
+describe("LibraryLayout", () => {
+  it("renders outlet", async () => {
     const { Route } = await import("./library");
-    const deps = { page: 1, pageSize: 50, sort: "title-asc" };
-    const result = await (Route.options.loader as unknown as (args: Record<string, unknown>) => Promise<unknown>)({ deps });
-    expect(getFilteredLibraryWorksServerFnMock).toHaveBeenCalledWith({ data: deps });
-    expect(getActiveJobCountServerFnMock).toHaveBeenCalled();
-    expect(getBulkReadingProgressServerFnMock).toHaveBeenCalled();
-    expect(result).toEqual({
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    });
-  });
-
-  it("renders 'Library' heading", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("Library")).toBeTruthy();
-  });
-
-  it("renders grid view by default", async () => {
-    mockView = "grid";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("The Great Gatsby", ["F. Scott Fitzgerald"], ["EBOOK"])],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByTestId("library-grid")).toBeTruthy();
-  });
-
-  it("renders table view with real DataTable when preference is table", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("The Great Gatsby", ["F. Scott Fitzgerald"], ["EBOOK"]), makeWork("Moby Dick", [], [])],
-        totalCount: 2,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("The Great Gatsby")).toBeTruthy();
-    expect(screen.getByText("Moby Dick")).toBeTruthy();
-    expect(screen.getByText("F. Scott Fitzgerald")).toBeTruthy();
-  });
-
-  it("renders title as link to work detail in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("The Great Gatsby", ["F. Scott Fitzgerald"], ["EBOOK"])],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const titleLink = screen.getByText("The Great Gatsby").closest("a");
-    expect(titleLink).toBeTruthy();
-    expect(titleLink?.getAttribute("href")).toBe("/library/work-the-great-gatsby");
-  });
-
-  it("renders work with no authors showing dash in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Unknown Author Book", [])],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("Unknown Author Book")).toBeTruthy();
-    expect(screen.getAllByText("—").length).toBeGreaterThan(0);
-  });
-
-  it("renders work with no editions showing dash for publisher/isbn in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [{
-          id: "work-no-editions",
-          titleDisplay: "No Editions",
-          titleCanonical: "no editions",
-          sortTitle: "no editions",
-          coverPath: null,
-          createdAt: new Date("2025-01-01"),
-          enrichmentStatus: "ENRICHED",
-          series: null,
-          seriesPosition: null,
-          editions: [],
-        }],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("No Editions")).toBeTruthy();
-  });
-
-  it("renders format badges in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Test Book", ["Author"], ["AUDIOBOOK"])],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("AUDIOBOOK")).toBeTruthy();
-  });
-
-  it("renders LibraryToolbar", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByTestId("library-toolbar")).toBeTruthy();
-  });
-
-  it("renders LibraryFilters", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByTestId("library-filters")).toBeTruthy();
-  });
-
-  it("renders LibraryPagination", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByTestId("library-pagination")).toBeTruthy();
-  });
-
-  it("shows empty state when no works and not scanning with no filters", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("No works yet")).toBeTruthy();
-    expect(screen.getByText("settings")).toBeTruthy();
-  });
-
-  it("empty state links to settings/libraries", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const link = screen.getByText("settings");
-    expect(link.getAttribute("href")).toBe("/settings/libraries");
-  });
-
-  it("does not show empty state when scanning with no works", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 1,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-    expect(screen.getByText(/Scanning/)).toBeTruthy();
-  });
-
-  it("does not show empty state when filters are active and results are empty", async () => {
-    mockSearch = { ...mockSearch, q: "something" } as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("shows scanning indicator when activeJobCount > 0", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 2,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText(/Scanning/)).toBeTruthy();
-  });
-
-  it("shows scanning indicator with new count when totalCount > prevCount", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Old Book")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    const { rerender } = render(<LibraryPage />);
-
-    await new Promise((resolve) => setTimeout(resolve, 0));
-
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Old Book"), makeWork("New Book")], totalCount: 2, facetCounts: defaultFacetCounts },
-      activeJobCount: 1,
-      progressMap: {},
-    };
-    rerender(<LibraryPage />);
-
-    expect(screen.getByText(/Scanning.*new/)).toBeTruthy();
-  });
-
-  it("shows processing badge for stub works in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Stub Book", ["Author"], ["EBOOK"], "STUB")],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.getByText("Processing\u2026")).toBeTruthy();
-  });
-
-  it("does not show processing badge for enriched works in table view", async () => {
-    mockView = "table";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Enriched Book", ["Author"], ["EBOOK"], "ENRICHED")],
-        totalCount: 1,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("Processing\u2026")).toBeNull();
-  });
-
-  it("passes progressMap to LibraryGrid", async () => {
-    mockView = "grid";
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: { "work-test": 42 },
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const grid = screen.getByTestId("library-grid");
-    expect(grid.getAttribute("data-progress-map")).toBe(JSON.stringify({ "work-test": 42 }));
-  });
-
-  it("does not show scanning indicator when activeJobCount is 0", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText(/Scanning/)).toBeNull();
-  });
-
-  it("navigates with search text when onSearchChange is called", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const onSearchChange = capturedToolbarProps.onSearchChange as (v: string) => void;
-    onSearchChange("hello");
-    expect(mockNavigate).toHaveBeenCalledWith(expect.objectContaining({
-      to: ".",
-      replace: true,
-    }));
-  });
-
-  it("navigates with empty q when search is cleared", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const onSearchChange = capturedToolbarProps.onSearchChange as (v: string) => void;
-    onSearchChange("");
-    expect(mockNavigate).toHaveBeenCalled();
-  });
-
-  it("navigates with sort when onSortChange is called", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const onSortChange = capturedToolbarProps.onSortChange as (v: string) => void;
-    onSortChange("title-desc");
-    expect(mockNavigate).toHaveBeenCalled();
-  });
-
-  it("navigates with filters when onFiltersChange is called", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const onFiltersChange = capturedFiltersProps.onFiltersChange as (v: Record<string, unknown>) => void;
-    onFiltersChange({ format: ["EBOOK"] });
-    expect(mockNavigate).toHaveBeenCalled();
-  });
-
-  it("navigates with page when onPageChange is called", async () => {
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 100, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const onPageChange = capturedPaginationProps.onPageChange as (v: number) => void;
-    onPageChange(2);
-    expect(mockNavigate).toHaveBeenCalled();
-  });
-
-  it("filters works by reading status", async () => {
-    mockView = "grid";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Unread"), makeWork("Reading"), makeWork("Done")],
-        totalCount: 3,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: { "work-reading": 50, "work-done": 100 },
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    const { rerender } = render(<LibraryPage />);
-
-    // Initially shows all
-    expect(screen.getByText("Grid: 3 works")).toBeTruthy();
-
-    // Switch to reading filter
-    const onFilterChange = capturedToolbarProps.onFilterChange as (v: string) => void;
-    onFilterChange("reading");
-    rerender(<LibraryPage />);
-    expect(screen.getByText("Grid: 1 works")).toBeTruthy();
-  });
-
-  it("shows finished works when reading filter is finished", async () => {
-    mockView = "grid";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Unread"), makeWork("Done")],
-        totalCount: 2,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: { "work-done": 100 },
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-
-    const onFilterChange = capturedToolbarProps.onFilterChange as (v: string) => void;
-    onFilterChange("finished");
-    const { rerender } = render(<LibraryPage />);
-    rerender(<LibraryPage />);
-    expect(screen.getAllByText(/Grid:/).length).toBeGreaterThan(0);
-  });
-
-  it("shows unread works when reading filter is unread", async () => {
-    mockView = "grid";
-    mockLoaderData = {
-      libraryResult: {
-        works: [makeWork("Unread"), makeWork("Done")],
-        totalCount: 2,
-        facetCounts: defaultFacetCounts,
-      },
-      activeJobCount: 0,
-      progressMap: { "work-done": 100 },
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-
-    const onFilterChange = capturedToolbarProps.onFilterChange as (v: string) => void;
-    onFilterChange("unread");
-    const { rerender } = render(<LibraryPage />);
-    rerender(<LibraryPage />);
-    expect(screen.getAllByText(/Grid:/).length).toBeGreaterThan(0);
-  });
-
-  it("does not show empty state when format filter is active", async () => {
-    mockSearch = { ...mockSearch, format: ["EBOOK"] } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("does not show empty state when authorId filter is active", async () => {
-    mockSearch = { ...mockSearch, authorId: ["a1"] } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("does not show empty state when seriesId filter is active", async () => {
-    mockSearch = { ...mockSearch, seriesId: ["s1"] } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("does not show empty state when publisher filter is active", async () => {
-    mockSearch = { ...mockSearch, publisher: ["Penguin"] } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("does not show empty state when hasCover filter is active", async () => {
-    mockSearch = { ...mockSearch, hasCover: true } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [], totalCount: 0, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(screen.queryByText("No works yet")).toBeNull();
-  });
-
-  it("passes search params as toolbar search value", async () => {
-    mockSearch = { ...mockSearch, q: "test query" } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    expect(capturedToolbarProps.searchValue).toBe("test query");
-  });
-
-  it("passes current filters from search to LibraryFilters", async () => {
-    mockSearch = { ...mockSearch, format: ["EBOOK"], hasCover: true } as unknown as typeof mockSearch;
-    mockLoaderData = {
-      libraryResult: { works: [makeWork("Test")], totalCount: 1, facetCounts: defaultFacetCounts },
-      activeJobCount: 0,
-      progressMap: {},
-    };
-    const { Route } = await import("./library");
-    const LibraryPage = Route.options.component as React.ComponentType;
-    render(<LibraryPage />);
-    const filters = capturedFiltersProps.filters as Record<string, unknown>;
-    expect(filters.format).toEqual(["EBOOK"]);
-    expect(filters.hasCover).toBe(true);
+    const LibraryLayout = (Route.options.component as React.ComponentType);
+    render(<LibraryLayout />);
+    expect(screen.getByTestId("outlet")).toBeTruthy();
   });
 });
